flash updater cannot work!!

+++ Platform/modules/rtclib/rtclib.c
@@ -119,7 +119,7 @@ void rtc_time_to_tm(unsigned long time, rtc_time *tm)
     tm->tm_mday = days;
 
     tm->tm_hour = time / 3600;
-    time -= tm->tm_hour * 3600;
+    time -= (unsigned long)tm->tm_hour * 3600;
     tm->tm_min = time / 60;
     tm->tm_sec = time - tm->tm_min * 60;
